+++ app/Http/Controllers/PvController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\Pv;
+use App\Models\Lot;
 use Illuminate\Http\Request;
+use App\Models\AssocPvDemande;
+use Illuminate\Support\Facades\Session;
+use RealRashid\SweetAlert\Facades\Alert;
 
 /**
  * Class PvController
@@ -19,9 +23,15 @@ class PvController extends Controller
     public function index()
     {
         $pvs = Pv::all();
-
-        return view('pv.index', compact('pvs'))
-            ;
+        if ($message = Session::get('error')) {
+            Alert::toast($message, 'error', '#fff')->position('bottom-end')->autoClose(10000)->timerProgressBar();
+            Session::remove('error');
+        }
+        if ($message = Session::get('success')) {
+            Alert::toast($message, 'success', '#fff')->position('bottom-end')->autoClose(10000)->timerProgressBar();
+            Session::remove('success');
+        }
+        return view('pv.index', compact('pvs'));
     }
 
     /**
@@ -48,7 +58,7 @@ class PvController extends Controller
         $pv = Pv::create($request->all());
 
         return redirect()->route('pv.index')
-            ->with('success', 'Pv created successfully.');
+            ->with('success', 'Pv Créé avec succès.');
     }
 
     /**
@@ -91,7 +101,7 @@ class PvController extends Controller
         $pv->update($request->all());
 
         return redirect()->route('pv.index')
-            ->with('success', 'Pv updated successfully');
+            ->with('success', 'Pv mis à jour avec succès');
     }
 
     /**
@@ -101,9 +111,28 @@ class PvController extends Controller
      */
     public function destroy($id)
     {
+        if (Lot::where('CodePV', $id)->count() != 0) {
+            return redirect()->route('pv.index')
+                ->with('error', 'Impossible de supprimer ce PV, Un ou plusieurs lots y sont encore associé (même si ces lots sont vide) ');
+        }
         $pv = Pv::find($id)->delete();
 
         return redirect()->route('pv.index')
-            ->with('success', 'Pv deleted successfully');
+            ->with('success', 'Pv supprimé avecc succes');
+    }
+
+    public function cloturerPV(int $codePV)
+    {
+        if (Lot::where('CodePV', $codePV)->where('status', 'OUVERT')->exists()) {
+            return back()->with('error', 'Ce PV Contient encore des lots OUVERT');
+        }
+        if (AssocPvDemande::where('CodePV', $codePV)->count() < 1) {
+            return back()->with('error', 'Ce PV ne contient aucun avis d\'aucun commissaire pour le moment, cloture impossible');
+        }
+        
+        Lot::where('CodePV', $codePV)->delete();
+        $pv = Pv::findOrFail($codePV);
+        $pv->update(['statut' => 'cloturé']);
+        return back()->with('success', 'PV :' . $pv->Refence_PV . ' Cloturé avec succès');
     }
 }
